Update stream processor from UI while connected (#14)
Overall I'm on the fence about this -- this seems a little too complicated and trying to handle too many cases. It might have been cleaner to just force the user to disconnect (and create an entirely new StreamProcessor object) if they want to change these parameters. That said, everything seems to work correctly.

Summary of changes:
- split "delimiter" LineEdit into two separate UI items, one for ascii (still a LineEdit) and one for binary (a SpinBox)
- depending on ascii/binary combobox, enable/disable appropriate UI items
- update stream processor to respond to changes in ascii/binary combobox, delimiter, binary message type, and number of ascii streams
- on pause, just start and stop timer, don't disconnect slot for updating plots
- on disconnect, stop timer for updating plots
- add additional test server on adjacent port, emitting binary data