Add ZMQ plumbing to both receive (ZMQ SUB: forward rcv'd data to serial) and emit serial data (ZMQ PUB) (#16)
It is often useful to allow multiple access on a serial port. This PR allows splot to provide multiple access via ZMQ. splot can publish serial data on a given local port, and can allow publishers to connect and send data to the serial device. This occurs via a separate thread that just busywaits for zmq messages, and if its receives any, dumps them into the serial.